requirements.txt
Flask
flask-cors
pymongo
python-dotenv
requests
gunicorn
